Improve handling of Alluxio test directory
Previously we made the dangerous assumption that only one
process would run the tests at a time per node. This was
violated by pull request builds running at the same time as
Alluxio master builds. This change removes the assumption by
only deleting files older than 2 hours.
